swap_management: Add command line options to start/stop zram swap
swap_management relies on D-Bus to start/stop zram swap during system
boottime/shutdown. However, since D-Bus behaves async, it would cause
race condition with other components. Using newly added command line
options to make sure swap_management starts/stops zram swap in sync with
upstart configs.

BUG=b:294122409
TEST=check /mnt/stateful_partition/shutdown_umount_failure.log, see if
swap_management is still activated during chromeos_shutdown.

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4751302
Auto-Submit: Chun-Tse Shao <ctshao@google.com>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Chun-Tse Shao <ctshao@google.com>

+++ swap_management/swap_tool.cc
@@ -3,7 +3,6 @@
 // found in the LICENSE file.
 
 #include "swap_management/swap_tool.h"
-#include "swap_management/swap_tool_status.h"
 
 #include <cinttypes>
 #include <utility>
@@ -195,9 +194,9 @@ absl::StatusOr<uint64_t> SwapTool::GetMemTotal() {
 // given in bips. A "bip" or "basis point" is 1/100 of 1%.
 absl::Status SwapTool::SetDefaultLowMemoryMargin(uint64_t mem_total) {
   // Calculate critical margin in MiB, which is 5.2% free. Ignore the decimal.
-  uint64_t critical_margin = (mem_total / 1024) * 0.052;
+  uint64_t critical_margin = (static_cast<double>(mem_total) / 1024) * 0.052;
   // Calculate moderate margin in MiB, which is 40% free. Ignore the decimal.
-  uint64_t moderate_margin = (mem_total / 1024) * 0.4;
+  uint64_t moderate_margin = (static_cast<double>(mem_total) / 1024) * 0.4;
   // Write into margin special file.
   return SwapToolUtil::Get()->WriteFile(
       base::FilePath("/sys/kernel/mm/chromeos-low_mem/margin"),
